fix: :bug: vertical scroll moved out to stop rerendering of whole page
